Move semantic root-order lowering into the compiler
